Split hemiMixTool
